feat: add voice channel music playback and restructure codebase into feature modules
  Music playback feature:
  - Add songbird integration for Discord voice channel audio
  - Implement /play and /stop slash commands for music control
  - Add fuzzy song search to find tracks by partial name
  - Support configurable music directory via MUSIC_DIR env var
  - Add music-related error types with user-friendly messages

  Codebase restructure:
  - Extract AI chatbot logic from bot.rs into chatbot/ module
    - chatbot/handler.rs: bot mention handling
    - chatbot/conversation.rs: reply chain history building
    - chatbot/context.rs: dynamic system prompt generation
    - chatbot/response.rs: Discord message sending
    - chatbot/tool_loop.rs: AI tool execution loop
  - Restructure auto_response.rs into auto_response/ module
    - auto_response/rules.rs: rule types and hardcoded rules
    - auto_response/matching.rs: fuzzy pattern matching logic
    - auto_response/handler.rs: auto-response event handler
  - Simplify bot.rs to thin event router (~140 lines)
  - Use named module files (e.g., chatbot.rs) instead of mod.rs

  The three distinct features now have clear separation:
  1. Auto-responses: fast pattern matching, no AI
  2. AI chatbot: mentions trigger full conversation with tools
  3. Music: slash commands for voice playback

  Documentation:
  - Update README with image/audio generation feature
  - Add MUSIC_DIR optional environment variable to setup
