+++ Cargo.toml
@@ -11,6 +11,7 @@ dotenvy = "0.15.7"
 env_logger = "0.11.8"
 eventsource-stream = "0.2.3"
 futures = "0.3.31"
+fuzzy-matcher = "0.3.7"
 hound = "3.5.1"
 log = "0.4.29"
 mime = "0.3.17"
@@ -23,8 +24,10 @@ reqwest = { version = "0.12.28", default-features = false, features = [
 ] }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
+songbird = { version = "0.5.0", features = ["builtin-queue"] }
 strsim = "0.11.1"
 strum = { version = "0.27.2", features = ["derive"] }
+symphonia = { version = "0.5.5", features = ["flac"] }
 thiserror = "2.0.18"
 tokio = { version = "1.49.0", features = ["full"] }
